tickets/src/routes/update.ts
import {
	NotAuthorizedError,
	NotFoundError,
	requireAuth,
	validateRequest,
} from '@zapro-msa/common';
import express, { Request, Response } from 'express';
import { body } from 'express-validator';
import { Ticket } from '../models/ticket';

const router = express.Router();

router.put(
	'/api/v1/tickets/:id',
	requireAuth,
	[
		body('title').not().isEmpty().withMessage('Title is required'),
		body('price')
			.isFloat({ gt: 0 })
			.withMessage('Price is required and must be greater than 0'),
	],
	validateRequest,
	async (req: Request, res: Response) => {
		const ticket = await Ticket.findById(req.params.id);

		if (!ticket) {
			throw new NotFoundError();
		}

		const { title, price } = req.body;

		if (ticket.userId !== req.currentUser!.id) {
			throw new NotAuthorizedError();
		}

		ticket.set({ title, price });

		await ticket.save();

		return res.send(ticket);
	}
);

export { router as updateTicketRouter };
